Move trace module into event_listener
