+++ evm_loader/lib/src/commands/trace.rs
@@ -1,13 +1,11 @@
+use crate::event_listener::trace::{TraceCallConfig, TraceConfig, TracedCall};
 use crate::{
     account_storage::EmulatorAccountStorage,
     commands::emulate::{emulate_trx, setup_syscall_stubs},
     errors::NeonError,
     event_listener::tracer::Tracer,
     rpc::Rpc,
-    types::{
-        trace::{TraceCallConfig, TraceConfig, TracedCall},
-        TxParams,
-    },
+    types::TxParams,
 };
 use evm_loader::types::Address;
 use serde::{Deserialize, Serialize};
