+++ evm_loader/lib/src/event_listener/trace.rs
